♻️  Refactor Models

+++ db/migrate/20210821185853_create_orders.rb
@@ -1,7 +1,7 @@
 class CreateOrders < ActiveRecord::Migration[6.1]
   def change
     create_table :orders do |t|
-      t.float :imei
+      t.decimal :imei
       t.float :annual_price, precision: 10, scale: 2
       t.string :device_model
       t.integer :installments
